Fixed Initial Simulation
I found the bug, I was not calculating the cross product correctly. Now the simulation should work. It is also commented and plots the position of the electron as it travels through the field. 

If you want to see how the strength of the field affects the electron, change line 103 to:
bx[infield] = 0.01 # for a weaker field
or
bx[infield] = 5 # for a stronger field